separate global score modifier compatibility test
